+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
 **./src/common/possible-route-tasks.js - countValidCompletedRoutes**
-* Echo divider to console after loop complete.
-* Separates console output for different test cases.
+* Removed console output.
+
+---
+
+**./submission.js - callPossibleRouteByDistanceTestCase**
+* Added 'case10' to 'resArr'